refactor: migrate output group
