+++ packages/webpack-cli/lib/utils/cli-flags.js
@@ -3,13 +3,11 @@ const cli = packageExists('webpack') ? require('webpack').cli : undefined;
 
 const HELP_GROUP = 'help';
 const BASIC_GROUP = 'basic';
-const OUTPUT_GROUP = 'output';
 const ADVANCED_GROUP = 'advanced';
 
 const groups = {
     HELP_GROUP,
     BASIC_GROUP,
-    OUTPUT_GROUP,
     ADVANCED_GROUP,
 };
 
@@ -55,7 +53,6 @@ const commands = [
             {
                 name: 'output',
                 type: String,
-                group: OUTPUT_GROUP,
                 description: 'To get the output in specified format ( accept json or markdown )',
             },
         ],
@@ -123,7 +120,6 @@ const core = [
         name: 'output',
         usage: '--output <path to output directory>',
         alias: 'o',
-        group: OUTPUT_GROUP,
         type: String,
         description: 'Output location of the file generated by webpack e.g. ./dist/',
         link: 'https://webpack.js.org/concepts/#output',
