+++ packages/webpack-cli/lib/utils/merge-strategies.js
@@ -1,0 +1,8 @@
+const outputStrategy = {
+    'output.filename': 'prepend',
+    'output.path': 'prepend',
+};
+
+module.exports = {
+    outputStrategy,
+};
